Fixed builds breaking on other computers because of incorrect file paths
I moved assets inside of game engine src folder and renamed it as well. The issue was that PackageCompiler also hardcodes in the development machine paths to grab assets. With the RelocatableFolders package, I use a constant for the assets folder. This brings the assets along with the build and updates the paths accordingly. In the future I would like to make it so a person making a game doesn't need to use RelocatableFolders themselves to specify file location