Add script functionality to entity editor
This commit adds the ability to add and delete scripts for entities in the editor. A new "Scripts" section has been added to the entity editor, which allows users to add and delete scripts associated with an entity. The user can input a name for each script using an input text field, and then click on "Add" or "Delete" buttons as needed.